requirements.txt
asgiref==3.7.2
Django==5.0.3
django-ckeditor==6.7.0
django-js-asset==2.2.0
pillow==10.2.0
python-decouple==3.8
sqlparse==0.4.4
tzdata==2024.1

